+++ sdk/storage/azure-storage-blobs/CHANGELOG.md
@@ -60,6 +60,7 @@
 - `MaxResults` in list APIs are renamed to `PageSizeHint`.
 - All date time related strings are now changed to `Azure::Core::DateTime` type.
 - Replace `std::pair<int64_t, int64_t>` with `Azure::Core::Http::Range` to denote blob ranges.
+- Move version strings into `Details` namespace.
 
 ## 12.0.0-beta.5 (2020-11-13)