+++ sdk/storage/azure-storage-files-shares/CHANGELOG.md
@@ -15,6 +15,7 @@
 - Replaced all transactional content MD5/CRC64 with `ContentHash` struct.
 - `FileShareHttpHeaders` is renamed to `ShareFileHttpHeaders`, and member `std::string ContentMd5` is changed to `Storage::ContentHash ContentHash`.
 - All date time related strings are now changed to `Azure::Core::DateTime` type.
+- Move version strings into `Details` namespace.
 
 ## 12.0.0-beta.5 (2020-11-13)